+++ src/components/rules/RuleCard.vue
@@ -1,8 +1,8 @@
 <template>
   <div class="card hover:bg-base-200 gap-2 p-2 text-sm">
-    <div class="min-h-6">
+    <div class="min-h-6 leading-6">
       <span>{{ index }}.</span>
-      <span class="ml-2 capitalize">{{ rule.type }}</span>
+      <span class="ml-2">{{ rule.type }}</span>
       <span
         class="text-main ml-2"
         v-if="rule.payload"
@@ -11,28 +11,35 @@
       </span>
       <span
         v-if="typeof size === 'number' && size !== -1"
-        class="badge badge-sm bg-base-200 ml-2"
+        class="text-base-content/80 ml-1 text-xs"
       >
-        {{ size }}
+        ({{ size }})
+        <QuestionMarkCircleIcon
+          v-if="size === 0"
+          class="ml-1 h-4 w-4"
+          @mouseenter="showMMDBSizeTip"
+        />
       </span>
       <button
         v-if="isUpdateableRuleSet"
-        :class="twMerge('btn btn-circle btn-xs ml-2', isUpdating ? 'animate-spin' : '')"
+        :class="
+          twMerge('btn btn-circle btn-ghost btn-xs -mt-1 ml-1', isUpdating ? 'animate-spin' : '')
+        "
         @click="updateRuleProviderClickHandler"
       >
         <ArrowPathIcon class="h-4 w-4" />
       </button>
     </div>
-    <div class="text-base-content/80 flex items-center gap-1">
+    <div class="flex items-center gap-2">
       <ProxyName
         :name="rule.proxy"
-        class="text-xs"
+        class="badge badge-sm gap-0"
       />
       <template v-if="proxyNode?.now && displayNowNodeInRule">
         <ArrowRightCircleIcon class="h-4 w-4" />
         <ProxyName
           :name="getNowProxyNodeName(rule.proxy)"
-          class="text-xs"
+          class="badge badge-sm gap-0"
         />
       </template>
       <span
@@ -51,13 +58,19 @@ import { updateRuleProviderAPI } from '@/api'
 import { useBounceOnVisible } from '@/composables/bouncein'
 import { NOT_CONNECTED } from '@/constant'
 import { getColorForLatency } from '@/helper'
+import { useTooltip } from '@/helper/tooltip'
 import { getLatencyByName, getNowProxyNodeName, proxyMap } from '@/store/proxies'
 import { fetchRules, ruleProviderList } from '@/store/rules'
 import { displayLatencyInRule, displayNowNodeInRule } from '@/store/settings'
 import type { Rule } from '@/types'
-import { ArrowPathIcon, ArrowRightCircleIcon } from '@heroicons/vue/24/outline'
+import {
+  ArrowPathIcon,
+  ArrowRightCircleIcon,
+  QuestionMarkCircleIcon,
+} from '@heroicons/vue/24/outline'
 import { twMerge } from 'tailwind-merge'
 import { computed, ref } from 'vue'
+import { useI18n } from 'vue-i18n'
 import ProxyName from '../proxies/ProxyName.vue'
 
 const props = defineProps<{
@@ -65,6 +78,8 @@ const props = defineProps<{
   index: number
 }>()
 
+const { t } = useI18n()
+const { showTip } = useTooltip()
 const proxyNode = computed(() => proxyMap.value[props.rule.proxy])
 const latency = computed(() => getLatencyByName(props.rule.proxy, props.rule.proxy))
 const latencyColor = computed(() => getColorForLatency(Number(latency.value)))
@@ -99,6 +114,10 @@ const updateRuleProviderClickHandler = async () => {
   await updateRuleProviderAPI(props.rule.payload)
   fetchRules()
   isUpdating.value = false
+}
+
+const showMMDBSizeTip = (e: Event) => {
+  showTip(e, t('mmdbSizeTip'))
 }
 
 useBounceOnVisible()