+++ src/constant/index.ts
@@ -7,6 +7,8 @@ import {
   SwatchIcon,
 } from '@heroicons/vue/24/outline'
 
+export const IS_APPLE_DEVICE = /Mac|iPod|iPhone|iPad/.test(navigator.platform)
+
 export const GLOBAL = 'GLOBAL'
 export const TEST_URL = 'https://www.gstatic.com/generate_204'
 export const IPV6_TEST_URL = 'https://ipv6.google.com/generate_204'
@@ -23,6 +25,11 @@ export enum FONTS {
   PING_FANG = 'PingFang',
   FIRA_SANS = 'FiraSans',
   SYSTEM_UI = 'SystemUI',
+}
+
+export enum EMOJIS {
+  TWEMOJI = 'twemoji',
+  NOTO_COLOR_EMOJI = 'noto-color-emoji',
 }
 
 export enum CONNECTIONS_TABLE_ACCESSOR_KEY {