+++ src/store/settings.ts
@@ -1,9 +1,11 @@
 import {
   CONNECTIONS_TABLE_ACCESSOR_KEY,
   DETAILED_CARD_STYLE,
+  EMOJIS,
   FONTS,
   GLOBAL,
   IP_INFO_API,
+  IS_APPLE_DEVICE,
   LANG,
   PROXY_CARD_SIZE,
   PROXY_CHAIN_DIRECTION,
@@ -52,7 +54,23 @@ export const isSidebarCollapsed = computed({
     isSidebarCollapsedConfig.value = value
   },
 })
-export const font = useStorage<FONTS>('config/font', FONTS.MI_SANS)
+const fontConfig = useStorage<FONTS>('config/font', FONTS.MI_SANS)
+export const font = computed({
+  get: () => {
+    const mode = import.meta.env.MODE
+    if (Object.values(FONTS).includes(mode as FONTS)) {
+      return mode as FONTS
+    }
+    return fontConfig.value
+  },
+  set: (val) => {
+    fontConfig.value = val
+  },
+})
+export const emoji = useStorage<EMOJIS>(
+  'config/emoji',
+  IS_APPLE_DEVICE ? EMOJIS.TWEMOJI : EMOJIS.NOTO_COLOR_EMOJI,
+)
 export const customBackgroundURL = useStorage('config/custom-background-image', '')
 export const dashboardTransparent = useStorage('config/dashboard-transparent', 90)
 export const autoUpgrade = useStorage('config/auto-upgrade', false)
@@ -85,6 +103,7 @@ export const displayProxiesRelationship = useStorage('config/display-proxies-rel
 
 // proxies
 export const collapseGroupMap = useStorage<Record<string, boolean>>('config/collapse-group-map', {})
+export const displayFinalOutbound = useStorage('config/show-seleted-for-now-node', false)
 export const twoColumnProxyGroup = useStorage('config/two-columns', true)
 export const speedtestUrl = useStorage<string>('config/speedtest-url', TEST_URL)
 export const independentLatencyTest = useStorage('config/independent-latency-test', false)